Add interface to work with graph

+++ src/mission_publisher/src/mission_publisher.cpp
@@ -147,7 +147,7 @@ MissionPublisher::Mission MissionPublisher::parseMissionFromJson(const nlohmann:
     const auto& velocities = mission_json["velocities"];
     for (const auto& v : velocities) {
 
-        mission.velocities.push_back(v.get<int>());
+        mission.velocities.push_back(v.get<float>());
     }
     
     return mission;
